requirements.txt
appnope==0.1.3
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
ipykernel==6.26.0
ipython==8.18.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tifffile==2023.9.26
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers @ git+https://github.com/huggingface/transformers.git@35551f9a0f66a22de4971b4a51b3c172d3b87f95
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
